IPObjGroup entity
