+++ src/database/Query.ts
@@ -5,6 +5,7 @@ import * as sqlstring from "sqlstring";
 export default class Query {
 
     public query(query:string, params: Array<any>, callback: (err:any, result: any) => void);
+    public query(query:string, params: {}, callback: (err:any, result: any) => void);
     public query(query: string, callback: (err: any, result: any) => void);
     public query(query: string, params: any = [] , callback?: (err: any, result: any) => void): void {
