+++ src/models/firewall/Firewall.ts
@@ -1003,9 +1003,9 @@ export class Firewall extends Model {
 						  	
 						  Verify too that these objects are not being used in any group.
 				*/
-				const r1 = await Interface.searchInterfaceUsageOutOfThisFirewall(req);
-				const r2 = await openvpnModel.searchOpenvpnUsageOutOfThisFirewall(req);
-				const r3 = await openvpnPrefixModel.searchPrefixUsageOutOfThisFirewall(req);
+				const r1: any = await Interface.searchInterfaceUsageOutOfThisFirewall(req);
+				const r2: any = await openvpnModel.searchOpenvpnUsageOutOfThisFirewall(req);
+				const r3: any = await openvpnPrefixModel.searchPrefixUsageOutOfThisFirewall(req);
 
 				if (r1) search.restrictions = utilsModel.mergeObj(search.restrictions, r1.restrictions);
 				if (r2) search.restrictions = utilsModel.mergeObj(search.restrictions, r2.restrictions);
