+++ src/models/policy/position.js
@@ -27,7 +27,7 @@ import db from '../../database/DatabaseService';
 //create object
 var policyPositionModel = {};
 const IpobjModel = require('../../models/ipobj/ipobj');
-const Ipobj_gModel = require('../../models/ipobj/group');
+import { IPObjGroup } from '../../models/ipobj/IPObjGroup';
 import { Interface } from '../../models/interface/Interface';
 const openvpnModel = require('../../models/vpn/openvpn/openvpn');
 const openvpnPrefixModel = require('../../models/vpn/openvpn/prefix');
@@ -232,7 +232,7 @@ policyPositionModel.getRulePositionData = position => {
 						if (item.ipobj>0 && item.type==='O') // IPOBJ
 						  data = await IpobjModel.getIpobj(dbCon, position.fwcloud, item.ipobj);
 						else if (item.ipobj_g>0 && item.type==='O') // IPOBJ GROUP
-							data = await Ipobj_gModel.getIpobj_g(dbCon, position.fwcloud, item.ipobj_g);
+							data = await IPObjGroup.getIpobj_g(dbCon, position.fwcloud, item.ipobj_g);
 						else if (item.interface>0 || item.type==='I') { // Network interface.
 							data = await Interface.getInterface(position.fwcloud, (item.type==='I')?item.ipobj:item.interface);
 							item.type='I'; // We need this when we create the data_policy_position_ipobjs object.
